Add to services check null and basic Unit Tests

+++ main/Services/Implementation/CustomProgramService.cs
@@ -13,18 +13,18 @@ public class CustomProgramService : ICustomProgramService
         _mapper = mapper;
     }
 
-    public async Task<IEnumerable<CustomProgramResponseDTO>> GetAllAsync()
+    public async Task<IEnumerable<CustomProgramResponseDTO>> GetAllAsync() // may returns void list
     {
         var programs = await _repository.GetAllAsync();
-
-        // if(programs == null) 
-
         return _mapper.Map<IEnumerable<CustomProgramResponseDTO>>(programs);
     }
 
     public async Task<CustomProgramResponseDTO> CreateAsync(CustomProgramCreateDTO dto, int creatorId)
     {
         var program = _mapper.Map<CustomProgram>(dto);
+        
+        if(program == null) throw new KeyNotFoundException($"Program not found");
+
         program.CreatorId = creatorId;
         await _repository.AddAsync(program);
         await _repository.SaveChangesAsync();
@@ -34,6 +34,9 @@ public class CustomProgramService : ICustomProgramService
     public async Task<CustomProgramResponseDTO?> GetByIdAsync(int id)
     {
         var program = await _repository.GetByIDAsync(id);
+
+        if(program == null) throw new KeyNotFoundException($"Program {id} not found");
+
         return _mapper.Map<CustomProgramResponseDTO?>(program);
     }
 
@@ -42,12 +45,19 @@ public class CustomProgramService : ICustomProgramService
         var program = _mapper.Map<CustomProgram>(dto);
         program.CustProgId = id;
         var updated = await _repository.UpdateAsync(program);
+
+        if(updated == null) throw new KeyNotFoundException("Updated program not found");
+
         await _repository.SaveChangesAsync(); // on every transaction
         return _mapper.Map<CustomProgramResponseDTO>(updated);
     }
 
     public async Task<bool> DeleteAsync(int id)
     {
+        var entity = await _repository.GetByIDAsync(id);
+
+        if(entity == null) throw new KeyNotFoundException("Not found");
+
         await _repository.DeleteByIDAsync(id);
         await _repository.SaveChangesAsync();
         return true;
