+++ main/Services/Implementation/WorkoutExerciseSetService.cs
@@ -1,6 +1,4 @@
-using Microsoft.EntityFrameworkCore;
 using AutoMapper;
-using Microsoft.AspNetCore.Http.Metadata;
 
 namespace FitnesTracker;
 
@@ -26,29 +24,41 @@ public class WorkoutExerciseService : IWorkoutExerciseService
     public async Task<WorkoutExerciseSetResponseDTO> GetSetByIdAsync(int id)
     {
         var entity = await _repository.GetByIDAsync(id);
+
+        if(entity == null) throw new KeyNotFoundException($"Not found");
+
         return _mapper.Map<WorkoutExerciseSetResponseDTO>(entity);
     }
 
     public async Task<IEnumerable<WorkoutExerciseSetResponseDTO>> GetSessionSetsAsync(int sessionId)
     {
-        var item = await _repository.GetByWorkoutSessionIdAsync(sessionId);
-        return _mapper.Map<IEnumerable<WorkoutExerciseSetResponseDTO>>(item);
+        var session = await _repository.GetByWorkoutSessionIdAsync(sessionId);
+
+        if(session == null) throw new KeyNotFoundException($"Session with id: {sessionId} not found");
+        
+        return _mapper.Map<IEnumerable<WorkoutExerciseSetResponseDTO>>(session);
     }
 
     public async Task<WorkoutExerciseSetResponseDTO> UpdateSetAsync(int id, WorkoutExerciseSetUpdateDTO dto)
     {
         var entity = await _repository.GetByIDAsync(id);
-        
+
+        if(entity == null) throw new KeyNotFoundException($"Exercise with id: {id} not found");
+
         entity.Weight = dto.Weight;
         entity.Reps = dto.Reps;
-        
+
         await _repository.UpdateAsync(entity);
         await _repository.SaveChangesAsync();
-        return _mapper.Map<WorkoutExerciseSetResponseDTO>(entity); // замени updated на entity
+        return _mapper.Map<WorkoutExerciseSetResponseDTO>(entity);
     }
 
     public async Task DeleteSetAsync(int id)
     {
+        var exercise = await _repository.GetByIDAsync(id);
+
+        if(exercise == null) throw new KeyNotFoundException($"Exercise with id: {id} not found");
+
         await _repository.DeleteByIDAsync(id);
         await _repository.SaveChangesAsync();
     }
